requirements.txt
Flask
google-genai
Flask-Cors
